fix compatbility with 64 bits and gcc 4.2

+++ src/Rcpp/traits/wrap_type_traits.h
@@ -54,7 +54,7 @@ template <> struct wrap_type_traits<int> { typedef wrap_type_primitive_tag wrap_
 template <> struct wrap_type_traits<double> { typedef wrap_type_primitive_tag wrap_category; } ;
 template <> struct wrap_type_traits<Rbyte> { typedef wrap_type_primitive_tag wrap_category; } ;
 template <> struct wrap_type_traits<Rcomplex> { typedef wrap_type_primitive_tag wrap_category; } ;
-template <> struct wrap_type_traits<size_t> { typedef wrap_type_primitive_tag wrap_category; } ;
+template <> struct wrap_type_traits<unsigned int> { typedef wrap_type_primitive_tag wrap_category; } ;
 template <> struct wrap_type_traits<bool> { typedef wrap_type_primitive_tag wrap_category; } ;
 template <> struct wrap_type_traits<std::string> { typedef wrap_type_primitive_tag wrap_category; } ;
 template <> struct wrap_type_traits<char> { typedef wrap_type_primitive_tag wrap_category; } ;
